move upload file javascript code to upload_script.js and show error message for no input
